Separate password validity check out
