+++ src/main/java/seedu/address/ui/LoginScreen.java
@@ -13,8 +13,8 @@ import javafx.scene.control.PasswordField;
 import javafx.stage.Stage;
 import seedu.address.MainApp;
 import seedu.address.commons.core.LogsCenter;
+import seedu.address.commons.util.PasswordUtil;
 import seedu.address.encryption.exceptions.UnsupportedPasswordException;
-import seedu.address.logic.parser.PasswordCommandParser;
 
 public class LoginScreen extends UiPart<Stage> {
     private static final String FXML = "LoginWindow.fxml";
@@ -63,7 +63,7 @@ public class LoginScreen extends UiPart<Stage> {
 
     private void handlePassword() {
         // blocks the case when / is appended to the end of password
-        if (!PasswordCommandParser.isValidPassword(userInputPassword.getText())) {
+        if (!PasswordUtil.isValidPassword(userInputPassword.getText())) {
             responseDisplay.setText("Wrong password, try again!");
             userInputPassword.clear();
             return;
